requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
supabase
PyJWT
pydantic[email]
python-multipart
bcrypt